Mac OS integration
Added stddef.h in the core include header. This allows all files to use size_t that I thought was included from stdint.h but no, Mac uses the one from stddef.h

Added an implementation of the get_a_ticks function, same as the linux one but keeping it separate for now.