Cleaning an fixes.
Added some missing semicolons on the ArrayList functions. Did not cause
a problem but we never know. And all the other functions have one so
its more consistent.

Converted the fsutils file to use the old c header for stdint. Removed
the using namespace std. This is to be more consistent because I rarely
use the c++ headers.