Replaced sanitize_path_slashes implementation
The documentation of the method states that it converts all slashes to
forward slashes. The windows version of the method was taking
frontslashes and turning them to two backslashes. A new method should
be created to convert slashes to a platform required format.

Fixed the implementation of get_file_extension too to return an empty
string if the path has no period.