Geometry upgrade
Added an FPolygon function to return a copy of the polygon positioned at the origin. This would make it easy to take a polygon and draw it on another texture since the vertices carry a position that might be out of bound on the texture.

Converted some vector method to return itself for easier chaining and to avoid creating duplicate vector instances that needs to be deleted.

Also fixed operator overloading on vectors by defining the overload outside of the class.