Corrected polygon offsetting
Turned all the arithmetic functions of a vector to return a value instead of a pointer. This was causing the current value to be calculated and returned. If we use value types everywhere we can overwrite the return values all we want.

Calculating the polygon offset every time we check the vertices array.